+++ CHANGES.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 1.133 - 2025-09-02
+
+### @cesium/engine
+
+#### Breaking Changes :mega:
+
+- Removed the argument fallback in `ITwinData.*` functions. Please switch to the new options argument signature [#12778](https://github.com/CesiumGS/cesium/issues/12778)
+
 ## 1.132 - 2025-08-01
 
 ### @cesium/engine
@@ -2454,6 +2462,7 @@ _This is an npm-only release to fix a publishing issue_.
     tileset.boundingSphere.center,
   );
   ```
+
   - This also fixes several issues with clipping planes not using the correct transform for tilesets with children.
 
 ### Additions :tada:
@@ -4299,6 +4308,7 @@ _This is an npm-only release to fix a publishing issue_.
                 isStopIncluded : true,
                 data : data
             });
+
     - `TimeInterval.fromIso8601` now takes a single options parameter. Code that looked like:
 
             TimeInterval.fromIso8601(intervalString, true, true, data);
@@ -4311,6 +4321,7 @@ _This is an npm-only release to fix a publishing issue_.
                 isStopIncluded : true,
                 data : data
             });
+
     - `interval.intersect(otherInterval)` -> `TimeInterval.intersect(interval, otherInterval)`
     - `interval.contains(date)` -> `TimeInterval.contains(interval, date)`